Split chapter and voice tests into focused files
