--- tests/test_api_voices.py
@@ -1,720 +1,0 @@
-import pytest
-import os
-import io
-import json
-import uuid
-import zipfile
-from pathlib import Path
-from unittest.mock import patch, MagicMock
-from app.db.core import init_db
-
-
-@pytest.fixture(autouse=True)
-def voices_root(tmp_path, monkeypatch):
-    import app.config
-    import app.web
-    import app.api.routers.voices
-    import app.api.routers.voices_helpers
-    import app.api.routers.voices_management
-    import app.api.routers.voices_bundles
-    import app.jobs.speaker
-
-    voices_dir = (tmp_path / "voices").resolve()
-    monkeypatch.setattr(app.web, "VOICES_DIR", voices_dir)
-    monkeypatch.setattr(app.config, "VOICES_DIR", voices_dir)
-    monkeypatch.setattr(app.api.routers.voices, "VOICES_DIR", voices_dir)
-    monkeypatch.setattr(app.api.routers.voices_helpers, "VOICES_DIR", voices_dir)
-    monkeypatch.setattr(app.jobs.speaker, "VOICES_DIR", voices_dir)
-    return voices_dir
-
-
-@pytest.fixture
-def client():
-    from fastapi.testclient import TestClient
-    from app.web import app as fastapi_app
-    return TestClient(fastapi_app)
-
-@pytest.fixture
-def clean_db(tmp_path):
-    from app.web import app as fastapi_app
-    db_path = "/tmp/test_api_voices.db"
-    if os.path.exists(db_path):
-        os.unlink(db_path)
-    os.environ["DB_PATH"] = db_path
-    import app.db.core
-    import importlib
-    importlib.reload(app.db.core)
-    init_db()
-
-    yield
-
-    if os.path.exists(db_path):
-        os.unlink(db_path)
-    fastapi_app.dependency_overrides = {}
-
-def test_list_speaker_profiles(clean_db, voices_root, client):
-    voices_dir = voices_root
-    voices_dir.mkdir()
-    (voices_dir / "SpeakerA").mkdir()
-    (voices_dir / "SpeakerA" / "voice.json").write_text(json.dumps({"version": 2, "name": "SpeakerA", "default_variant": "Default"}))
-    (voices_dir / "SpeakerA" / "Default").mkdir()
-    (voices_dir / "SpeakerA" / "Default" / "profile.json").write_text(json.dumps({"variant_name": "Default"}))
-    (voices_dir / "SpeakerA" / "Default" / "v1.wav").write_text("audio")
-
-    with patch("app.jobs.get_speaker_settings", return_value={"built_samples": [], "speed": 1.0, "test_text": "", "engine": "xtts"}):
-        response = client.get("/api/speaker-profiles")
-        assert response.status_code == 200
-        data = response.json()
-        assert len(data) == 1
-        assert data[0]["name"] == "SpeakerA"
-        assert data[0]["engine"] == "xtts"
-        assert data[0]["asset_base_url"] == "/out/voices/SpeakerA/Default"
-
-
-def _write_voice_root(root: Path, voice_name: str = "Dracula") -> Path:
-    voice_root = root / voice_name
-    default_dir = voice_root / "Default"
-    angry_dir = voice_root / "Angry"
-    default_dir.mkdir(parents=True)
-    angry_dir.mkdir(parents=True)
-    (voice_root / "voice.json").write_text(json.dumps({
-        "version": 2,
-        "name": voice_name,
-        "default_variant": "Default",
-    }))
-    (default_dir / "profile.json").write_text(json.dumps({
-        "variant_name": "Default",
-        "engine": "xtts",
-    }))
-    (default_dir / "latent.pth").write_bytes(b"latent")
-    (default_dir / "sample.mp3").write_bytes(b"mp3")
-    (default_dir / "sample.wav").write_bytes(b"preview")
-    (default_dir / "source.wav").write_bytes(b"source")
-    (angry_dir / "profile.json").write_text(json.dumps({
-        "variant_name": "Angry",
-        "engine": "xtts",
-    }))
-    (angry_dir / "source.wav").write_bytes(b"angry-source")
-    return voice_root
-
-
-def _zip_names(response) -> set[str]:
-    with zipfile.ZipFile(io.BytesIO(response.content)) as zf:
-        return set(zf.namelist())
-
-
-def test_export_voice_bundle_excludes_source_wavs_by_default(clean_db, voices_root, client):
-    voices_root.mkdir()
-    _write_voice_root(voices_root)
-
-    response = client.get("/api/voices/Dracula/bundle/download")
-
-    assert response.status_code == 200
-    names = _zip_names(response)
-    assert "bundle.json" in names
-    assert "voice.json" in names
-    assert "Default/profile.json" in names
-    assert "Default/latent.pth" in names
-    assert "Default/sample.mp3" in names
-    assert "Default/sample.wav" in names
-    assert "Default/source.wav" not in names
-    assert "Angry/source.wav" not in names
-
-
-def test_export_voice_bundle_includes_source_wavs_when_requested(clean_db, voices_root, client):
-    voices_root.mkdir()
-    _write_voice_root(voices_root)
-
-    response = client.get("/api/voices/Dracula/bundle/download?include_source_wavs=true")
-
-    assert response.status_code == 200
-    names = _zip_names(response)
-    assert "Default/source.wav" in names
-    assert "Angry/source.wav" in names
-
-
-def test_import_voice_bundle_duplicate_creates_suffixed_copy(clean_db, voices_root, client):
-    voices_root.mkdir()
-    _write_voice_root(voices_root)
-    export_response = client.get("/api/voices/Dracula/bundle/download?include_source_wavs=true")
-
-    response = client.post(
-        "/api/voices/bundle/import",
-        files={"file": ("dracula.voice.zip", io.BytesIO(export_response.content), "application/zip")},
-    )
-
-    assert response.status_code == 200
-    payload = response.json()
-    assert payload["voice_name"] == "Dracula 2"
-    assert payload["was_renamed"] is True
-    assert (voices_root / "Dracula").exists()
-    assert (voices_root / "Dracula 2" / "Default" / "latent.pth").exists()
-    assert (voices_root / "Dracula 2" / "Angry" / "source.wav").exists()
-    imported_manifest = json.loads((voices_root / "Dracula 2" / "voice.json").read_text())
-    assert imported_manifest["name"] == "Dracula 2"
-    assert imported_manifest["default_variant"] == "Default"
-
-
-def test_import_voice_bundle_rejects_invalid_archives(clean_db, voices_root, client):
-    voices_root.mkdir()
-
-    malformed = client.post(
-        "/api/voices/bundle/import",
-        files={"file": ("bad.zip", io.BytesIO(b"not a zip"), "application/zip")},
-    )
-    assert malformed.status_code == 400
-
-    missing_voice = io.BytesIO()
-    with zipfile.ZipFile(missing_voice, "w") as zf:
-        zf.writestr("Default/profile.json", "{}")
-    response = client.post(
-        "/api/voices/bundle/import",
-        files={"file": ("missing.voice.zip", io.BytesIO(missing_voice.getvalue()), "application/zip")},
-    )
-    assert response.status_code == 400
-    assert "voice.json" in response.json()["message"]
-
-    traversal = io.BytesIO()
-    with zipfile.ZipFile(traversal, "w") as zf:
-        zf.writestr("voice.json", json.dumps({"name": "Evil"}))
-        zf.writestr("../escape.txt", "nope")
-        zf.writestr("Default/profile.json", "{}")
-    response = client.post(
-        "/api/voices/bundle/import",
-        files={"file": ("traversal.voice.zip", io.BytesIO(traversal.getvalue()), "application/zip")},
-    )
-    assert response.status_code == 400
-
-    missing_profile = io.BytesIO()
-    with zipfile.ZipFile(missing_profile, "w") as zf:
-        zf.writestr("voice.json", json.dumps({"name": "NoProfile"}))
-        zf.writestr("Default/sample.wav", b"preview")
-    response = client.post(
-        "/api/voices/bundle/import",
-        files={"file": ("missing-profile.voice.zip", io.BytesIO(missing_profile.getvalue()), "application/zip")},
-    )
-    assert response.status_code == 400
-    assert "profile.json" in response.json()["message"]
-
-
-def test_imported_latent_voice_lists_ready_without_rebuild(clean_db, voices_root, client):
-    source_root = voices_root / "source"
-    source_root.mkdir(parents=True)
-    _write_voice_root(source_root, "LatentVoice")
-
-    from app.domain.voices.bundles import export_voice_bundle
-    bundle = export_voice_bundle(source_root, "LatentVoice")
-
-    response = client.post(
-        "/api/voices/bundle/import",
-        files={"file": ("latent.voice.zip", io.BytesIO(bundle), "application/zip")},
-    )
-    assert response.status_code == 200
-
-    bridge = MagicMock()
-    bridge.check_readiness.return_value = (True, "ready")
-    with patch("app.engines.bridge.create_voice_bridge", return_value=bridge):
-        profiles = client.get("/api/speaker-profiles").json()
-
-    profile = next(p for p in profiles if p["name"] == "LatentVoice")
-    assert profile["has_latent"] is True
-    assert profile["is_ready"] is True
-    assert profile["is_rebuild_required"] is False
-
-
-def test_legacy_profile_listing_repairs_missing_speaker_rows_and_preserves_default_switch(clean_db, voices_root, client):
-    voices_root.mkdir()
-    legacy_speaker_id = str(uuid.uuid4())
-
-    for profile_name, variant_name in [
-        ("Dark Fantasy - Default", "Default"),
-        ("Dark Fantasy - Light Narrator", "Light Narrator"),
-    ]:
-        profile_dir = voices_root / profile_name
-        profile_dir.mkdir()
-        (profile_dir / "profile.json").write_text(json.dumps({
-            "speaker_id": legacy_speaker_id,
-            "variant_name": variant_name,
-            "engine": "xtts",
-        }))
-
-    profiles_response = client.get("/api/speaker-profiles")
-    assert profiles_response.status_code == 200
-    profiles = profiles_response.json()
-
-    repaired_default = next(p for p in profiles if p["name"] == "Dark Fantasy - Default")
-    repaired_variant = next(p for p in profiles if p["name"] == "Dark Fantasy - Light Narrator")
-    assert repaired_default["speaker_id"] == legacy_speaker_id
-    assert repaired_variant["speaker_id"] == legacy_speaker_id
-
-    speakers_response = client.get("/api/speakers")
-    assert speakers_response.status_code == 200
-    repaired_speaker = next(s for s in speakers_response.json() if s["id"] == legacy_speaker_id)
-    assert repaired_speaker["name"] == "Dark Fantasy"
-    assert repaired_speaker["default_profile_name"] == "Dark Fantasy - Default"
-
-    set_default_response = client.post("/api/settings/default-speaker", data={"name": "Dark Fantasy - Default"})
-    assert set_default_response.status_code == 200
-
-    refreshed_profiles = client.get("/api/speaker-profiles").json()
-    refreshed_default = next(p for p in refreshed_profiles if p["name"] == "Dark Fantasy - Default")
-    assert refreshed_default["is_default"] is True
-
-def test_create_and_delete_profile(clean_db, voices_root, client):
-    voices_dir = voices_root
-    voices_dir.mkdir()
-
-    # Create
-    response = client.post("/api/speaker-profiles", data={"speaker_id": "S1", "variant_name": "V1"})
-    assert response.status_code == 200
-    name = response.json()["name"]
-    # Verify nested structure
-    if " - " in name:
-        spk, var = name.split(" - ", 1)
-        assert (voices_dir / spk.strip() / var.strip()).exists()
-    else:
-        assert (voices_dir / name).exists()
-
-    # Delete
-    response = client.delete(f"/api/speaker-profiles/{name}")
-    assert response.status_code == 200
-    if " - " in name:
-        spk, var = name.split(" - ", 1)
-        assert not (voices_dir / spk.strip() / var.strip()).exists()
-    else:
-        assert not (voices_dir / name).exists()
-
-
-def test_character_voice_assignment_blank_value_clears_to_default(clean_db, voices_root, client):
-    from app.db.projects import create_project
-
-    voices_root.mkdir()
-    pid = create_project("Character Project", "/tmp")
-
-    create_response = client.post(f"/api/projects/{pid}/characters", data={"name": "Alice", "speaker_profile_name": "Prof1"})
-    assert create_response.status_code == 200
-    character_id = create_response.json()["character_id"]
-
-    update_response = client.put(f"/api/characters/{character_id}", data={"speaker_profile_name": ""})
-    assert update_response.status_code == 200
-
-    chars_response = client.get(f"/api/projects/{pid}/characters")
-    assert chars_response.status_code == 200
-    characters = chars_response.json()["characters"]
-    assert characters[0]["speaker_profile_name"] is None
-
-
-def test_create_character_blank_voice_uses_default(clean_db, voices_root, client):
-    from app.db.projects import create_project
-
-    voices_root.mkdir()
-    pid = create_project("Character Project", "/tmp")
-
-    create_response = client.post(f"/api/projects/{pid}/characters", data={"name": "Bob", "speaker_profile_name": ""})
-    assert create_response.status_code == 200
-
-    chars_response = client.get(f"/api/projects/{pid}/characters")
-    assert chars_response.status_code == 200
-    characters = chars_response.json()["characters"]
-    assert characters[0]["speaker_profile_name"] is None
-
-
-def test_create_profile_persists_engine_metadata(clean_db, voices_root, client):
-    voices_root.mkdir()
-
-    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=True):
-        response = client.post("/api/speaker-profiles", data={"speaker_id": "S1", "variant_name": "Vox", "engine": "voxtral"})
-    assert response.status_code == 200
-
-    name = response.json()["name"]
-    spk, var = name.split(" - ", 1)
-    meta = json.loads((voices_root / spk.strip() / var.strip() / "profile.json").read_text())
-    assert meta["engine"] == "voxtral"
-
-
-def test_create_voxtral_profile_requires_api_key(clean_db, voices_root, client, monkeypatch):
-    voices_root.mkdir()
-
-    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=False):
-        response = client.post("/api/speaker-profiles", data={"speaker_id": "S1", "variant_name": "Vox", "engine": "voxtral"})
-    assert response.status_code == 400
-    assert "not enabled" in response.json()["message"]
-
-
-def test_update_profile_engine(clean_db, voices_root, client):
-    voices_root.mkdir()
-    profile_dir = voices_root / "SpeakerA"
-    profile_dir.mkdir()
-    (profile_dir / "profile.json").write_text(json.dumps({"variant_name": "Default"}))
-
-    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=True):
-        response = client.post("/api/speaker-profiles/SpeakerA/engine", data={"engine": "voxtral"})
-    assert response.status_code == 200
-    assert response.json()["engine"] == "voxtral"
-
-    meta = json.loads((profile_dir / "profile.json").read_text())
-    assert meta["engine"] == "voxtral"
-
-
-def test_update_voxtral_engine_requires_api_key(clean_db, voices_root, client, monkeypatch):
-    voices_root.mkdir()
-    profile_dir = voices_root / "SpeakerA"
-    profile_dir.mkdir()
-    (profile_dir / "profile.json").write_text(json.dumps({"variant_name": "Default", "engine": "xtts"}))
-
-    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=False):
-        response = client.post("/api/speaker-profiles/SpeakerA/engine", data={"engine": "voxtral"})
-    assert response.status_code == 400
-    assert "not enabled" in response.json()["message"]
-
-
-def test_update_profile_engine_rejects_invalid_value(clean_db, voices_root, client):
-    voices_root.mkdir()
-    profile_dir = voices_root / "SpeakerA"
-    profile_dir.mkdir()
-    (profile_dir / "profile.json").write_text(json.dumps({"variant_name": "Default"}))
-
-    response = client.post("/api/speaker-profiles/SpeakerA/engine", data={"engine": "bad-engine"})
-    assert response.status_code == 400
-
-
-def test_update_profile_reference_sample(clean_db, voices_root, client):
-    voices_root.mkdir()
-    profile_dir = voices_root / "SpeakerA"
-    profile_dir.mkdir()
-    (profile_dir / "profile.json").write_text(json.dumps({"variant_name": "Default", "engine": "voxtral"}))
-    (profile_dir / "sample1.wav").write_text("audio")
-
-    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=True):
-        response = client.post("/api/speaker-profiles/SpeakerA/reference-sample", data={"sample_name": "sample1.wav"})
-    assert response.status_code == 200
-    assert response.json()["reference_sample"] == "sample1.wav"
-
-    meta = json.loads((profile_dir / "profile.json").read_text())
-    assert meta["reference_sample"] == "sample1.wav"
-
-
-def test_update_profile_voxtral_voice_id(clean_db, voices_root, client):
-    voices_root.mkdir()
-    profile_dir = voices_root / "SpeakerA"
-    profile_dir.mkdir()
-    (profile_dir / "profile.json").write_text(json.dumps({"variant_name": "Default", "engine": "voxtral"}))
-
-    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=True):
-        response = client.post("/api/speaker-profiles/SpeakerA/voxtral-voice-id", data={"voice_id": "voice_123"})
-    assert response.status_code == 200
-    assert response.json()["voxtral_voice_id"] == "voice_123"
-
-    meta = json.loads((profile_dir / "profile.json").read_text())
-    assert meta["voxtral_voice_id"] == "voice_123"
-
-
-def test_voxtral_profile_test_accepts_saved_voice_id_without_samples(clean_db, voices_root, client):
-    profile_dir = voices_root / "SpeakerA"
-    profile_dir.mkdir(parents=True)
-    (profile_dir / "profile.json").write_text(json.dumps({
-        "variant_name": "Default",
-        "engine": "voxtral",
-        "voxtral_voice_id": "voice_123",
-    }))
-
-    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=True), patch("app.state.put_job"), patch("app.jobs.enqueue"):
-        response = client.post("/api/speaker-profiles/SpeakerA/test")
-
-    assert response.status_code == 200
-    assert response.json()["status"] == "ok"
-
-
-def test_voice_test_job_uses_descriptive_queue_title(clean_db, voices_root, client):
-    profile_dir = voices_root / "Alice - Warm"
-    profile_dir.mkdir(parents=True)
-    (profile_dir / "1.wav").write_text("audio")
-    (profile_dir / "profile.json").write_text(json.dumps({
-        "speaker_id": "speaker-alice",
-        "variant_name": "Warm",
-        "engine": "xtts",
-    }))
-
-    with patch("app.state.put_job") as mock_put_job, patch("app.jobs.enqueue"):
-        response = client.post("/api/speaker-profiles/Alice%20-%20Warm/test")
-
-    assert response.status_code == 200
-    queued_job = mock_put_job.call_args.args[0]
-    assert queued_job.custom_title == "Building voice for Alice: Warm"
-
-def test_character_crud(clean_db, client):
-    from app.db.projects import create_project
-    pid = create_project("P1")
-
-    # Create
-    response = client.post(f"/api/projects/{pid}/characters", data={"name": "Alice", "speaker_profile_name": "Prof1"})
-    assert response.status_code == 200
-    cid = response.json()["id"]
-
-    # List
-    response = client.get(f"/api/projects/{pid}/characters")
-    assert response.status_code == 200
-    assert len(response.json()["characters"]) == 1
-
-    # Update
-    response = client.put(f"/api/characters/{cid}", data={"name": "Alice Updated", "color": "#123456"})
-    assert response.status_code == 200
-
-    # Delete
-    response = client.delete(f"/api/characters/{cid}")
-    assert response.status_code == 200
-
-def test_speaker_crud(clean_db, client):
-    # Create
-    response = client.post("/api/speakers", data={"name": "Narrator", "default_profile_name": "P1"})
-    assert response.status_code == 200
-    sid = response.json()["id"]
-
-    # List
-    response = client.get("/api/speakers")
-    assert response.status_code == 200
-    assert len(response.json()) == 1
-
-    # Update
-    response = client.put(f"/api/speakers/{sid}", data={"name": "Narrator Updated"})
-    assert response.status_code == 200
-
-    # Delete
-    response = client.delete(f"/api/speakers/{sid}")
-    assert response.status_code == 200
-
-def test_rename_profile_and_security(clean_db, voices_root, client):
-    voices_dir = voices_root
-    voices_dir.mkdir()
-    (voices_dir / "OldName").mkdir()
-    (voices_dir / "OldName" / "profile.json").write_text(json.dumps({"variant_name": "Old"}))
-    (voices_dir / "OldName" / "latent.pth").write_text("latent")
-
-    # Success
-    response = client.post("/api/voices/rename-profile", data={"old_name": "OldName", "new_name": "NewName - Variant"})
-    assert response.status_code == 200
-    assert (voices_dir / "NewName" / "Variant").exists()
-    assert not (voices_dir / "OldName").exists()
-    assert (voices_dir / "NewName" / "Variant" / "latent.pth").exists()
-
-    # Security traversal
-    response = client.post("/api/voices/rename-profile", data={"old_name": "NewName - Variant", "new_name": "../../traversal"})
-    assert response.status_code == 403
-
-def test_speaker_settings_updates(clean_db, client):
-    with patch("app.jobs.update_speaker_settings") as mock_update:
-        # Test text
-        response = client.post("/api/speaker-profiles/SpeakerA/test-text", data={"text": "Hello world"})
-        assert response.status_code == 200
-        assert response.json()["test_text"] == "Hello world"
-
-        # Speed
-        response = client.post("/api/speaker-profiles/SpeakerA/speed", data={"speed": 1.2})
-        assert response.status_code == 200
-        assert response.json()["speed"] == 1.2
-
-    assert mock_update.call_count == 2
-
-def test_reset_speaker_test_text(clean_db, voices_root, client):
-    from app.api.routers.voices import DEFAULT_SPEAKER_TEST_TEXT
-
-    voices_dir = voices_root
-    profile_dir = voices_dir / "SpeakerA"
-    profile_dir.mkdir(parents=True)
-    (profile_dir / "profile.json").write_text(json.dumps({"test_text": "Custom text"}))
-
-    with patch("app.jobs.speaker.VOICES_DIR", voices_dir):
-        response = client.post("/api/speaker-profiles/SpeakerA/reset-test-text")
-
-    assert response.status_code == 200
-    assert response.json()["test_text"] == DEFAULT_SPEAKER_TEST_TEXT
-    # Check that test_text was removed from profile.json (falling back to default)
-    # SpeakerA is a flat folder in this test setup
-    assert "test_text" not in json.loads((profile_dir / "profile.json").read_text())
-
-def test_build_and_test_profiles(clean_db, voices_root, client):
-    voices_dir = voices_root
-    voices_dir.mkdir()
-
-    with patch("app.api.routers.voices_helpers.VOICES_DIR", voices_dir):
-        # Build
-        file_content = b"fake wav"
-        files = {"files": ("input1.wav", io.BytesIO(file_content), "audio/wav")}
-        (voices_dir / "SpeakerA").mkdir()
-        (voices_dir / "SpeakerA" / "profile.json").write_text("{}")
-        with patch("app.state.put_job"), patch("app.jobs.enqueue"):
-            response = client.post("/api/speaker-profiles/SpeakerA/build", files=files)
-            assert response.status_code == 200
-            # Build creates SpeakerA/Default for a flat name if it's missing
-            assert (voices_dir / "SpeakerA" / "input1.wav").exists() or (voices_dir / "SpeakerA" / "Default" / "input1.wav").exists()
-
-        (voices_dir / "SpeakerA" / "1.wav").write_text("fake wav content 2")
-
-        # Test
-        with patch("app.state.put_job"), patch("app.jobs.enqueue"):
-            response = client.post("/api/speaker-profiles/SpeakerA/test")
-            assert response.status_code == 200
-
-def test_build_and_test_require_samples(clean_db, voices_root, client):
-    voices_dir = voices_root
-    voices_dir.mkdir()
-
-    with patch("app.api.routers.voices_helpers.VOICES_DIR", voices_dir):
-        profile_dir = voices_dir / "SpeakerA"
-        profile_dir.mkdir(parents=True, exist_ok=True)
-
-        with patch("app.state.put_job"), patch("app.jobs.enqueue"):
-            response = client.post("/api/speaker-profiles/SpeakerA/test")
-            assert response.status_code == 400
-            assert "sample" in response.json()["message"].lower()
-
-            response = client.post("/api/speaker-profiles/SpeakerA/build")
-            assert response.status_code == 400
-            assert "sample" in response.json()["message"].lower()
-
-
-def test_xtts_voice_actions_reject_when_disabled(clean_db, voices_root, client):
-    voices_dir = voices_root
-    profile_dir = voices_dir / "SpeakerA"
-    profile_dir.mkdir(parents=True, exist_ok=True)
-    (profile_dir / "profile.json").write_text(json.dumps({
-        "variant_name": "Default",
-        "engine": "xtts",
-    }))
-    (profile_dir / "1.wav").write_text("fake wav content")
-
-    with patch("app.api.routers.voices_helpers._is_engine_active", return_value=False):
-        response = client.post("/api/speaker-profiles/SpeakerA/test")
-
-    assert response.status_code == 400
-    assert "enabled in Settings" in response.json()["message"]
-
-def test_legacy_build_and_rename(clean_db, voices_root, client):
-    voices_dir = voices_root
-    voices_dir.mkdir()
-
-    # Legacy Build
-    with patch("app.state.put_job"), patch("app.jobs.enqueue"):
-        response = client.post("/api/speaker-profiles/SpeakerL/build", files={"files": ("sample.wav", io.BytesIO(b"legacy"), "audio/wav")})
-        assert response.status_code == 200
-
-    # Legacy Rename
-    if not (voices_dir / "SpeakerL").exists():
-        (voices_dir / "SpeakerL").mkdir()
-    response = client.post("/api/speaker-profiles/SpeakerL/rename", data={"new_name": "SpeakerNew"})
-    assert response.status_code == 200
-
-def test_profile_creation_errors(clean_db, voices_root, client):
-    voices_dir = voices_root
-    voices_dir.mkdir()
-
-    # Already exists
-    (voices_dir / "S1 - V1").mkdir()
-    (voices_dir / "S1 - V1" / "profile.json").write_text("{}")
-    response = client.post("/api/speaker-profiles", data={"speaker_id": "S1", "variant_name": "V1"})
-    assert response.status_code == 400
-
-    # Traversal
-    response = client.post("/api/speaker-profiles", data={"speaker_id": "../../etc", "variant_name": "passwd"})
-    assert response.status_code == 403
-
-    # Exception
-    with patch("app.api.routers.voices_management.Path.mkdir", side_effect=Exception("boom")):
-        response = client.post("/api/speaker-profiles", data={"speaker_id": "Err", "variant_name": "V1"})
-        assert response.status_code == 500
-
-    response = client.post("/api/speaker-profiles", data={"speaker_id": "Err", "variant_name": "V1", "engine": "bad-engine"})
-    assert response.status_code == 400
-
-def test_rename_speaker_with_variants(clean_db, voices_root, client):
-    voices_dir = voices_root
-    voices_dir.mkdir()
-    (voices_dir / "Narrator").mkdir()
-    (voices_dir / "Narrator" / "Calm").mkdir()
-    (voices_dir / "Narrator" / "voice.json").write_text(json.dumps({"version": 2, "name": "Narrator"}))
-    (voices_dir / "Narrator" / "Calm" / "profile.json").write_text(json.dumps({"speaker_id": "Narrator"}))
-    (voices_dir / "Narrator" / "Calm" / "latent.pth").write_text("latent")
-    (voices_dir / "Narrator" / "Excited").mkdir()
-    (voices_dir / "Narrator" / "Excited" / "profile.json").write_text("{}")
-
-    # Rename
-    response = client.post("/api/voices/rename-profile", data={"old_name": "Narrator", "new_name": "Dracula"})
-    assert response.status_code == 200
-
-    # Verify both main and variants are renamed
-    assert (voices_dir / "Dracula").exists()
-    assert (voices_dir / "Dracula" / "Calm").exists()
-    assert (voices_dir / "Dracula" / "Excited").exists()
-    assert not (voices_dir / "Narrator").exists()
-    assert (voices_dir / "Dracula" / "Calm" / "latent.pth").exists()
-
-    # Verify metadata update
-    meta = json.loads((voices_dir / "Dracula" / "Calm" / "profile.json").read_text())
-    assert meta["speaker_id"] == "Dracula"
-
-def test_rename_profile_default_sync(clean_db, voices_root, client):
-    from app.state import update_settings, get_settings
-    update_settings({"default_speaker_profile": "OldProfile"})
-
-    voices_dir = voices_root
-    voices_dir.mkdir()
-    (voices_dir / "OldProfile").mkdir()
-    response = client.post("/api/voices/rename-profile", data={"old_name": "OldProfile", "new_name": "NewProfile"})
-    assert response.status_code == 200
-    assert get_settings()["default_speaker_profile"] == "NewProfile"
-
-def test_upload_samples_security_and_failure(clean_db, voices_root, client):
-    voices_dir = voices_root
-    voices_dir.mkdir()
-
-    # Define files for the test
-    files = {"files": ("sample.wav", io.BytesIO(b"data"), "audio/wav")}
-
-    # Exception during upload
-    with patch("app.api.routers.voices_actions.Path.mkdir", side_effect=Exception("makedirs failed")):
-        response = client.post("/api/speaker-profiles/SpeakerA/samples/upload", files=files)
-        assert response.status_code == 500
-
-def test_delete_sample_errors(clean_db, voices_root, client):
-    voices_dir = voices_root
-    voices_dir.mkdir()
-    (voices_dir / "SpeakerA").mkdir()
-    (voices_dir / "SpeakerA" / "profile.json").write_text("{}")
-    (voices_dir / "SpeakerA" / "sample1.wav").write_text("audio")
-
-    # Success
-    response = client.delete("/api/speaker-profiles/SpeakerA/samples/sample1.wav")
-    assert response.status_code == 200
-    assert not (voices_dir / "SpeakerA" / "sample1.wav").exists()
-
-    # Exception
-    with patch("pathlib.Path.unlink", side_effect=Exception("unlink failed")):
-        (voices_dir / "SpeakerA" / "bad.wav").write_text("trash")
-        response = client.delete("/api/speaker-profiles/SpeakerA/samples/bad.wav")
-        assert response.status_code == 500
-
-
-def test_delete_sample_rejects_traversal(voices_root):
-    from app.api.routers.voices import delete_speaker_sample
-
-    voices_dir = voices_root
-    voices_dir.mkdir()
-    (voices_dir / "SpeakerA").mkdir()
-    (voices_dir / "SpeakerA" / "profile.json").write_text("{}")
-    with patch("app.api.routers.voices_helpers.VOICES_DIR", voices_dir):
-        response = delete_speaker_sample(
-            name="SpeakerA",
-            sample_name="../../escape.wav",
-        )
-    assert response.status_code == 403
-
-def test_assign_profile_to_speaker_errors(clean_db, voices_root, client):
-    voices_dir = voices_root
-    voices_dir.mkdir()
-    (voices_dir / "SomeProf").mkdir()
-    (voices_dir / "SomeProf" / "profile.json").write_text("{}")
-
-    # Generic error
-    with patch("app.api.routers.voices_management.db.get_speaker", side_effect=Exception("db crash")):
-        response = client.post("/api/speaker-profiles/SomeProf/assign", data={"speaker_id": "sid"})
-        assert response.status_code == 500
